feat: add pyproject and requirements update for python 3.13
