+++ requirements.txt
@@ -19,11 +19,10 @@ requests
 ruff
 scikit-image
 scipy
-sentencepiece
 tensorboard
 timm>=0.8
 torch>=1.12
 torchvision>=0.13
 tqdm
-transformers==4.37.2
+transformers==4.53.2
 yapf
