--- setup.cfg
@@ -1,18 +1,0 @@
-[flake8]
-ignore =
-    # line break before binary operator (W503)
-    W503,
-    # line break after binary operator (W504)
-    W504,
-max-line-length=120
-
-[codespell]
-skip = .git,./docs/build,*.cfg,./pyiqa/archs/uranker.py,scripts/*
-quiet-level = 3
-ignore-words-list = hist,fo,NiN 
-
-[yapf]
-based_on_style = pep8
-column_limit = 120
-blank_line_before_nested_class_or_def = true
-split_before_expression_after_opening_paren = true
